New Feature: Fluid Pipes
WIP

+++ src/main/java/micdoodle8/mods/galacticraft/api/item/IItemOxygenSupply.java
@@ -7,7 +7,7 @@ public interface IItemOxygenSupply
     /*
      * Returns the amount of gas that this oxygen item is able to supply
      */
-	public float discharge(ItemStack itemStack, float amount);
+	public int discharge(ItemStack itemStack, int amount);
 
     public int getOxygenStored(ItemStack theItem);
 }
