+++ src/main/java/micdoodle8/mods/galacticraft/api/transmission/tile/IOxygenStorage.java
@@ -8,15 +8,15 @@ public interface IOxygenStorage
     /**
      * Sets the amount of oxygen this unit has stored.
      */
-    public void setOxygenStored(float oxygen);
+    public void setOxygenStored(int oxygen);
 
     /**
      * * @return Get the amount of oxygen currently stored in the block.
      */
-    public float getOxygenStored();
+    public int getOxygenStored();
 
     /**
      * @return Get the max amount of oxygen that can be stored in the block.
      */
-    public float getMaxOxygenStored();
+    public int getMaxOxygenStored();
 }
